+++ client/src/components/card/card.component.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
-import Card from '@material-ui/core/Card';
-import CardContent from '@material-ui/core/CardContent';
-import TextField from '@material-ui/core/TextField';
+import{ Card,CardContent,TextField} from '@material-ui/core';
 import { withStyles } from '@material-ui/core/styles';
 import styles from './card.styles';
