+++ client/src/components/setting-icon-btn/setting-icon-btn.styles.jsx
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    backgroundColor: '#3bb7e3',
+    width: 50,
+    height: 50,
+  },
+}));
+
+export default useStyles;
