+++ client/src/components/swiper/slides-data.js
@@ -1,0 +1,25 @@
+const slides = [
+  {
+    id: 1,
+    text: 'Welcome To Monster Analyzer',
+    subText: 'Anaylze your Network',
+  },
+  {
+    id: 2,
+    text: 'Select Your Network \n',
+    subText: 'Set Your Network to get the required information',
+  },
+  {
+    id: 3,
+    text: 'Customize Your Network \n',
+    subText: 'Edit your network to get the proper reports',
+  },
+  {
+    id: 4,
+    text: 'Find your hosts \n',
+    subText:
+      'Find information on hosts, computers, phones found on your local network',
+  },
+];
+
+export default slides
