+++ client/src/components/table/table.component.jsx
@@ -1,46 +1,18 @@
 import React from 'react';
-import { withStyles, makeStyles } from '@material-ui/core/styles';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
-import useStyles from './table.styles';
+import {
+  Table,
+  TableBody,
+  TableContainer,
+  TableHead,
+  TableRow,
+  Paper,
+} from '@material-ui/core';
+import { useStyles, StyledTableRow, StyledTableCell } from './table.styles';
 import SettingIconBtn from '../setting-icon-btn/setting-icon-btn';
-
-const StyledTableCell = withStyles((theme) => ({
-  head: {
-    backgroundColor: '#F5F5F5',
-    color: '#3BB7E3',
-    fontSize: 15,
-  },
-  body: {
-    fontSize: 14,
-  },
-}))(TableCell);
-
-const StyledTableRow = withStyles((theme) => ({
-  root: {
-    '&:nth-of-type(odd)': {
-      backgroundColor: theme.palette.action.hover,
-    },
-  },
-}))(TableRow);
-
-function createData(name, data) {
-  return { name, data };
-}
-
-const rows = [
-  createData('Network Interface', 'wlan0'),
-  createData('Filter', 'tcp or udp'),
-  createData('Gateway', '192.168.1.1'),
-];
 
 export default function DataTable({ hidden, handleClick, networkSetting }) {
   const classes = useStyles();
+
 
   return (
     <TableContainer
@@ -48,7 +20,7 @@ export default function DataTable({ hidden, handleClick, networkSetting }) {
       component={Paper}
       className={classes.container}
     >
-      <Table className={classes.table} aria-label="customized table">
+      <Table aria-label="customized table">
         <TableHead>
           <TableRow>
             <StyledTableCell>Your Customized Network Setting</StyledTableCell>
