+++ client/src/components/table/table.styles.jsx
@@ -1,9 +1,8 @@
 import { withStyles, makeStyles } from '@material-ui/core/styles';
+import TableCell from '@material-ui/core/TableCell';
+import TableRow from '@material-ui/core/TableRow';
 
 const useStyles = makeStyles({
-  table: {
-    // minWidth: 700,
-  },
   container: {
     width: 600,
     display: 'flex',
@@ -12,5 +11,23 @@ const useStyles = makeStyles({
     alignItems: 'center',
   },
 });
+const StyledTableCell = withStyles((theme) => ({
+  head: {
+    backgroundColor: '#F5F5F5',
+    color: '#3BB7E3',
+    fontSize: 15,
+  },
+  body: {
+    fontSize: 14,
+  },
+}))(TableCell);
 
-export default useStyles;
+const StyledTableRow = withStyles((theme) => ({
+  root: {
+    '&:nth-of-type(odd)': {
+      backgroundColor: theme.palette.action.hover,
+    },
+  },
+}))(TableRow);
+
+export { useStyles, StyledTableRow, StyledTableCell };
